Tres-Explorer/requirements.txt
Django==1.9.5
django-cors-headers==1.1.0
django-jsonfield==0.9.19
django-sslserver==0.19
djangorestframework==3.3.3
geopy==1.11.0
gunicorn==19.4.5
lxml==3.6.0
MySQL-python==1.2.5
oauthlib==1.0.3
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
wheel==0.24.0
whitenoise==3.0
